+++ resources/views/components/admin/form.blade.php
@@ -1,0 +1,3 @@
+{!! Form::model($model, compact('url', 'method')) !!}
+    {{ $slot }}
+{!! Form::close() !!}